Remove --start flag from call to hoogle
When a search is performed with :Hoogle, you should be able to press
the enter key on a function and have the temporary buffer display some
more detailed info about that particular function. However, instead of
doing that, instead the message 'Unknown flag: --start' is shown. This
commit removes the --start flag responsible for this error message.